+++ spec/features/maths_and_physics_claim_spec.rb
@@ -187,7 +187,13 @@ RSpec.feature "Maths & Physics claims" do
       expect(page).to have_text("Check your answers before sending your application")
 
       stub_geckoboard_dataset_update
-      stub_qualified_teaching_status_show(claim: claim)
+
+      stub_qualified_teaching_statuses_show(
+        query: {
+          trn: claim.teacher_reference_number,
+          ni: claim.national_insurance_number
+        }
+      )
 
       freeze_time do
         perform_enqueued_jobs do