+++ spec/features/student_loans_claim_spec.rb
@@ -189,7 +189,12 @@ RSpec.feature "Teacher Student Loan Repayments claims" do
 
       expect(page).to have_text("Check your answers before sending your application")
 
-      stub_qualified_teaching_status_show(claim: claim)
+      stub_qualified_teaching_statuses_show(
+        query: {
+          trn: claim.teacher_reference_number,
+          ni: claim.national_insurance_number
+        }
+      )
 
       freeze_time do
         perform_enqueued_jobs do