+++ spec/lib/dqt/api_spec.rb
@@ -4,7 +4,7 @@ module Dqt
   describe Api do
     subject(:api) { described_class.new(client: double("client")) }
 
-    it { should delegate_method(:qualified_teaching_status).to(:v1) }
+    it { should delegate_method(:qualified_teaching_statuses).to(:v1) }
 
     describe "#v1" do
       it "returns V1" do